+++ Kernel/Modules/CustomerTicketSearch.pm
@@ -610,7 +610,12 @@ sub Run {
                     }
                     else
                     {
+                        ARTICLE:
                         for my $Articles (@Article) {
+
+                            # Skip internal articles.
+                            next ARTICLE if $Articles->{ArticleType} =~ /-int/;
+
                             if ( $Articles->{Body} ) {
                                 $Data{ArticleTree}
                                     .= "\n-->||$Articles->{ArticleType}||$Articles->{From}||"
@@ -741,10 +746,9 @@ sub Run {
         elsif ( $GetParam{ResultForm} eq 'Print' ) {
             for my $TicketID (@ViewableTicketIDs) {
 
-                # get first article data
-                my %Article = $Self->{TicketObject}->ArticleLastCustomerArticle(
+                # Get last customer or any other article if it doesn't exist.
+                my %Article = $Self->_LastCustomerArticle(
                     TicketID      => $TicketID,
-                    Extended      => 1,
                     DynamicFields => 0,
                 );
 
@@ -920,10 +924,9 @@ sub Run {
                     )
                 {
 
-                    # get first article data
-                    my %Article = $Self->{TicketObject}->ArticleLastCustomerArticle(
+                    # Get last customer or any other article if it doesn't exist.
+                    my %Article = $Self->_LastCustomerArticle(
                         TicketID      => $TicketID,
-                        Extended      => 1,
                         DynamicFields => 1,
                     );
 
@@ -1695,6 +1698,47 @@ sub _StopWordsServerErrorsGet {
     }
 
     return %StopWordsServerErrors;
+}
+
+sub _LastCustomerArticle {
+    my ( $Self, %Param ) = @_;
+
+    # Get all customer articles.
+    my @Index = $Self->{TicketObject}->ArticleIndex(
+        TicketID   => $Param{TicketID},
+        SenderType => 'customer',
+    );
+
+    # Go over articles in reverse order and return the last external one.
+    if (@Index) {
+        for my $CustomerArticleID ( reverse @Index ) {
+            my %LastCustomerArticle = $Self->{TicketObject}->ArticleGet(
+                ArticleID     => $CustomerArticleID,
+                Extended      => 1,
+                DynamicFields => $Param{DynamicFields},
+            );
+            if ( $LastCustomerArticle{ArticleType} !~ /-int/ ) {
+                return %LastCustomerArticle;
+            }
+        }
+    }
+
+    # If no customer articles were found, return the last external one.
+    @Index = $Self->{TicketObject}->ArticleIndex(
+        TicketID => $Param{TicketID},
+    );
+    for my $ArticleID ( reverse @Index ) {
+        my %LastArticle = $Self->{TicketObject}->ArticleGet(
+            ArticleID     => $ArticleID,
+            Extended      => 1,
+            DynamicFields => $Param{DynamicFields},
+        );
+        if ( $LastArticle{StateType} eq 'merged' || $LastArticle{ArticleType} !~ /-int/ ) {
+            return %LastArticle;
+        }
+    }
+
+    return;
 }
 
 1;